完成门店与连锁关联，修改部分连锁部分接口

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/domain/store/StoreChainDTO.java
@@ -11,8 +11,8 @@ import lombok.*;
 @Builder
 public class StoreChainDTO extends SuperEntity {
     @TableField(value = "`store_id`")
-    private Integer  storeId;
+    private Long  storeId;
     @TableField(value = "`chain_id`")
-    private Integer  chainId;
+    private Long  chainId;
 
 }
